[inductor] Remove identity from ops.scan
Currently scan has an `init` argument which must be the identity of the
combine function. This isn't strictly necessary if we are more careful about
keeping track of the first element and avoid combining it with anything.

This does additionally require that there are no active load masks, since we can't
do the `where_cond` any more. However, this shouldn't be possible anyway since
scans are always realized and only fused via the scheduler.

ghstack-source-id: c4e33cd5221027918fd00208dd2567e580aa8c86
Pull Request resolved: https://github.com/pytorch/pytorch/pull/119727